W edge as variant of wobbly edge
